Update edges.js
Moved connect and destroy into the MakeEdge class. Destroy will need to be a static function, and both will need to have an array passed to them that they can operate upon. If that works, I will have the setup() method have an array be passed in as well, that way we can in theory run two triangulations simultaneously with their own separate arrays, AND I can get rid of the Edges class.